+++ ironflow-cli/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.8](https://gitlab.com/ThomasTartrau/ironflow/compare/ironflow-cli-v0.1.7...ironflow-cli-v0.1.8) - 2026-07-29
+
+### Added
+
+- #20 add secret, api-key, user and audit-log CLI commands
+
 ## [0.1.5](https://gitlab.com/ThomasTartrau/ironflow/compare/ironflow-cli-v0.1.4...ironflow-cli-v0.1.5) - 2026-07-27
 
 ### Added